Use flexbox instead to for the landing page
I changed the previous implementation of the landing page, which was
using a Grid that layouts elements horizontally/rows. The problem
with this is that it's hard to center the contents (`Grid` children)
because the layout is in horizontal--forcing you to use either padding
or margin for centering. Now, I use flexbox that layouts the content
in column so that `justify-content: "center"` easily centers them.